Refactor module method override
- Remove overriding entire request_builder module and keep just the new
version of add_facet_fq_to_solr in tul_cdm_helper.
- [TBD] It probably should go into a custom lib file.
