+++ app/helpers/tul_cdm_helper.rb
@@ -569,3 +569,32 @@ module TulCdmHelper
 	end
 
 end
+
+##
+# Override Blacklight bug which will not render sidebar facets in alphabetical order
+#
+module Blacklight::RequestBuilders
+  ##
+  # Add any existing facet limits, stored in app-level HTTP query
+  # as :f, to solr as appropriate :fq query. 
+  def add_facet_fq_to_solr(solr_parameters, user_params)   
+    # Force blacklgiht to render facets in alphabetical order
+    solr_parameters["facet.sort"] = "index"
+
+    # convert a String value into an Array
+    if solr_parameters[:fq].is_a? String
+      solr_parameters[:fq] = [solr_parameters[:fq]]
+    end
+
+    # :fq, map from :f. 
+    if ( user_params[:f])
+      f_request_params = user_params[:f] 
+      
+      f_request_params.each_pair do |facet_field, value_list|
+        Array(value_list).each do |value| next if value.blank? # skip empty strings
+          solr_parameters.append_filter_query facet_value_to_fq_string(facet_field, value)
+        end              
+      end      
+    end
+  end
+end
